fixed-graph-py/pyproject.toml
[tool.poetry]
name = "fixed-computational-graph"
version = "0.1.0"
description="👾 Fixed Computational Graph Example 👾."
authors=["bt3gl"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12.0"
python-dotenv = "0.21.0"


[tool.poetry.group.dev.dependencies]
tox = "^4.0.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
